move SessionDetailsView out of hqadmin
All views inside hqadmin are audited by auditcare, so every auth request
from formplayer makes 2 couchdb saves. This change should take 80-200ms off of
every request to formplayer
