+++ corehq/apps/hqadmin/urls.py
@@ -28,8 +28,8 @@ from corehq.apps.hqadmin.views import (
     system_ajax,
     pillow_operation_api,
     web_user_lookup,
-    top_five_projects_by_country,
-    SessionDetailsView)
+    top_five_projects_by_country)
+
 from corehq.apps.reports.dispatcher import AdminReportDispatcher
 
 from corehq.apps.api.urls import admin_urlpatterns as admin_api_urlpatterns
@@ -73,6 +73,5 @@ urlpatterns = [
         name=ReprocessMessagingCaseUpdatesView.urlname),
     url(r'^top_five_projects_by_country/$', top_five_projects_by_country, name='top_five_projects_by_country'),
     url(r'^web_user_data', WebUserDataView.as_view(), name=WebUserDataView.urlname),
-    url(r'^session_details/', SessionDetailsView.as_view(), name=SessionDetailsView.urlname),
     AdminReportDispatcher.url_pattern(),
 ]
